重たいライブラリを遅延ロードに変更
