--- src/py-recognition/src/filter_torch.py
@@ -1,4 +1,0 @@
-import numpy
-from src.filter import VoiceActivityDetectorFilter
-
-# いまはなにもない
